requirements.txt
Django==1.9.4
djangorestframework==3.3.2
requests
